requirements.txt
# --- Web Framework ---
Flask==2.3.3
Werkzeug==2.3.7

# --- Environment ---
python-dotenv==1.0.1

# --- OCR & Image Processing ---
opencv-python==4.9.0.80
pytesseract==0.3.10
easyocr==1.7.1

# --- AI & NLP ---
google-generativeai==0.3.2
sympy==1.12
num2words==0.5.12
gTTS==2.5.1
pyttsx3==2.90  # ✅ Fixed here

# --- PyTorch (for EasyOCR) ---
torch==2.2.1
torchvision==0.17.1
torchaudio==2.2.1

# --- Utilities ---
numpy==1.26.4
pytest==8.1.1
gunicorn
